# Conflicts:
#	src/main/java/org/ddolib/ddo/examples/pdp/DPDIncrementalHop.java
#	src/main/java/org/ddolib/ddo/examples/tsp/SmallestIncidentHopIncremental.java
#	src/main/java/org/ddolib/ddo/examples/tsp/TSPMain.java
#	src/main/java/org/ddolib/ddo/examples/tsp/TSPProblem.java
#	src/main/java/org/ddolib/ddo/examples/tsp/TSPRelax.java
#	src/main/java/org/ddolib/ddo/examples/tsp/TSPState.java

+++ src/main/java/org/ddolib/ddo/core/Problem.java
@@ -13,8 +13,10 @@ public interface Problem<T> {
     int nbVars();
     /** @return the intial state of the problem */
     T initialState();
-    /** @return the problem's initial value */
-    int initialValue();
+    /**
+     * @return the problem's initial value
+     */
+    double initialValue();
 
     /**
      * @param state the state from which the transitions should be applicable
@@ -31,12 +33,12 @@ public interface Problem<T> {
      * @param decision the decision which is applied to `state`. 
      */
     T transition(final T state, final Decision decision);
-    /** 
+    /**
      * Computes the impact on the objective value of making the given
      * decision in the specified state.
-     * 
-     * @param state the state from which the transition originates
-     * @param decision the decision which is applied to `state`. 
+     *
+     * @param state    the state from which the transition originates
+     * @param decision the decision which is applied to `state`.
      */
-    int transitionCost(final T state, final Decision decision);
+    double transitionCost(final T state, final Decision decision);
 }
